Change the font size estimation to mimic macvim
In a vim vertical spacing adjustment is done on integer values of pixels. Also
the font baseline in adjusted to center the font inside the new line height.
To match the adjustment of vertical spacing, horizontal spacing was changed to
integer values as well.

The method for calculating line height, width and baseline comes from a discussion
on the macvim mailing list[1] a few years ago. It is this the current method.

[1]: http://tech.groups.yahoo.com/group/vim-mac/message/4631